refactor: split up router by endpoint
